boards: nordic: nrf54h20: disable sec ipc
The secure domain IPC node statuses are being used to generate UICRs
which tell the secure domain to attempt connecting to the app and
radio cores. secure domain should only be enabled if used, so the
nodes should default to disabled.

Without this change, the H20 is unable to go to sleep since secure
domain stays awake, persistanly trying to connect to both cores.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>